user_management/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
beautifulsoup4==4.13.4
behave==1.2.6
blinker==1.7.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==43.0.3
cssselect==1.3.0
defusedxml==0.7.1
distlib==0.4.0
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
execnet==2.1.1
fastapi==0.104.1
fasteners==0.19
filelock==3.18.0
future==1.0.0
greenlet==3.2.3
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.27.0
hyperframe==6.0.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kaitaistruct==0.10
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mycdp==1.2.0
nodeenv==1.9.1
oic==1.7.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.7.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pyjwkest==1.4.2
PyJWT==2.9.0
pynose==1.5.4
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.1
pytest-xdist==3.8.0
python-dotenv==1.0.0
python-xlib==0.33
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
sbvirtualdisplay==1.4.0
selenium==4.34.2
selenium-wire==5.1.0
seleniumbase==4.40.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
starlette==0.27.0
tabcompleter==1.4.0
testcontainers==4.12.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
virtualenv==20.33.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
zstandard==0.23.0
